+++ src/Socket.cpp
@@ -771,7 +771,7 @@ int Socket::IsDataThere(int mills)
 {
 	fd_set rfds;
 	struct timeval tv;
-	int n, ret;
+	size_t n, ret;
 
 	try {
 		FD_ZERO(&rfds);
